Splited components and css functionalities
