+++ src/components/UserCard/UserCard.jsx
@@ -1,4 +1,5 @@
 /* eslint-disable react/prop-types */
+import './UserCard.css'
 
 const UserCard = ({userData}) => {
     return(
