+++ src/components/UserCardList/UserCatdList.css
@@ -1,0 +1,3 @@
+.user-cards-container{
+    display: flex;
+}
